rename events file
